[chore] data 패키지명 수정
