+++ app/src/main/java/com/linkedlist/linkllet/MainViewModel.kt
@@ -3,7 +3,7 @@ package com.linkedlist.linkllet
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.linkedlist.linkellet.core.data.repository.AuthRepository
+import com.linkedlist.linkllet.core.data.repository.AuthRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
