--- core/data/src/main/java/com/linkedlist/linkellet/core/data/RetrofitFactory.kt
@@ -1,35 +1,0 @@
-package com.linkedlist.linkellet.core.data
-
-import android.content.Context
-import com.linkedlist.linkellet.core.data.source.remote.api.KeyInterceptor
-import com.linkedlist.linkllet.core.data.BuildConfig
-import dagger.hilt.android.qualifiers.ApplicationContext
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object RetrofitFactory {
-
-    inline fun <reified Service> create(
-        context: Context
-    ): Service {
-        return Retrofit.Builder()
-            .baseUrl(BuildConfig.BACKEND_URL)
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(
-                OkHttpClient.Builder()
-                    .addInterceptor(createLoggingInterceptor())
-                    .addInterceptor(KeyInterceptor(context))
-                    .build()
-            )
-            .build()
-            .create(Service::class.java)
-    }
-
-    fun createLoggingInterceptor(): HttpLoggingInterceptor {
-        return HttpLoggingInterceptor().apply {
-            level = HttpLoggingInterceptor.Level.BODY
-        }
-    }
-}
