--- core/data/src/main/java/com/linkedlist/linkellet/core/data/di/NetworkModule.kt
@@ -1,31 +1,0 @@
-package com.linkedlist.linkellet.core.data.di
-
-import android.content.Context
-import com.linkedlist.linkellet.core.data.RetrofitFactory
-import com.linkedlist.linkellet.core.data.source.remote.api.AuthService
-import com.linkedlist.linkellet.core.data.source.remote.api.LinkService
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkModule {
-
-    @Provides
-    @Singleton
-    fun provideAuthService(
-        @ApplicationContext context : Context
-    ): AuthService
-        = RetrofitFactory.create(context)
-
-    @Provides
-    @Singleton
-    fun provideLinkService(
-        @ApplicationContext context : Context
-    ): LinkService
-        = RetrofitFactory.create(context)
-}
