--- core/data/src/main/java/com/linkedlist/linkellet/core/data/model/Auth.kt
@@ -1,5 +1,0 @@
-package com.linkedlist.linkellet.core.data.model
-
-enum class Auth {
-    SIGNED_UP, ALREADY_SIGNED_UP, FAILED
-}
