--- core/data/src/main/java/com/linkedlist/linkellet/core/data/model/Folder.kt
@@ -1,11 +1,0 @@
-package com.linkedlist.linkellet.core.data.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Folder(
-    val id: Long,
-    val name: String,
-    val type: String,
-    val size: Int,
-)
