--- core/data/src/main/java/com/linkedlist/linkellet/core/data/model/Link.kt
@@ -1,11 +1,0 @@
-package com.linkedlist.linkellet.core.data.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Link(
-    val id : Long,
-    val name : String,
-    val url : String,
-    val createAt : String
-)
