--- core/data/src/main/java/com/linkedlist/linkellet/core/data/model/request/AddLinkRequest.kt
@@ -1,9 +1,0 @@
-package com.linkedlist.linkellet.core.data.model.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AddLinkRequest(
-    val name : String,
-    val url : String
-)
