--- core/data/src/main/java/com/linkedlist/linkellet/core/data/model/request/SignUpRequest.kt
@@ -1,8 +1,0 @@
-package com.linkedlist.linkellet.core.data.model.request
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SignUpRequest(
-    val deviceId : String
-)
