--- core/data/src/main/java/com/linkedlist/linkellet/core/data/model/response/FolderResponse.kt
@@ -1,9 +1,0 @@
-package com.linkedlist.linkellet.core.data.model.response
-
-import com.linkedlist.linkellet.core.data.model.Folder
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class FolderResponse(
-    val folderList : List<Folder>
-)
