--- core/data/src/main/java/com/linkedlist/linkellet/core/data/model/response/LinkResponse.kt
@@ -1,9 +1,0 @@
-package com.linkedlist.linkellet.core.data.model.response
-
-import com.linkedlist.linkellet.core.data.model.Link
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class LinkResponse(
-    val articleList : List<Link>
-)
