--- core/data/src/main/java/com/linkedlist/linkellet/core/data/repository/AuthRepository.kt
@@ -1,7 +1,0 @@
-package com.linkedlist.linkellet.core.data.repository
-
-import kotlinx.coroutines.flow.Flow
-
-interface AuthRepository {
-    fun signUp() : Flow<Result<Boolean>>
-}
