--- core/data/src/main/java/com/linkedlist/linkellet/core/data/source/remote/AuthRemoteDataSource.kt
@@ -1,7 +1,0 @@
-package com.linkedlist.linkellet.core.data.source.remote
-
-import com.linkedlist.linkellet.core.data.model.Auth
-
-interface AuthRemoteDataSource {
-    suspend fun signUp() : Result<Auth>
-}
