+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/model/Auth.kt
@@ -1,0 +1,5 @@
+package com.linkedlist.linkllet.core.data.model
+
+enum class Auth {
+    SIGNED_UP, ALREADY_SIGNED_UP, FAILED
+}
