+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/model/Folder.kt
@@ -1,0 +1,11 @@
+package com.linkedlist.linkllet.core.data.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Folder(
+    val id: Long,
+    val name: String,
+    val type: String,
+    val size: Int,
+)
