+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/model/Link.kt
@@ -1,0 +1,11 @@
+package com.linkedlist.linkllet.core.data.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class Link(
+    val id : Long,
+    val name : String,
+    val url : String,
+    val createAt : String
+)
