+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/model/request/AddFolderRequest.kt
@@ -1,0 +1,8 @@
+package com.linkedlist.linkllet.core.data.model.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AddFolderRequest(
+    val name: String
+)
