+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/model/request/AddLinkRequest.kt
@@ -1,0 +1,9 @@
+package com.linkedlist.linkllet.core.data.model.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AddLinkRequest(
+    val name : String,
+    val url : String
+)
