+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/model/request/SignUpRequest.kt
@@ -1,0 +1,8 @@
+package com.linkedlist.linkllet.core.data.model.request
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SignUpRequest(
+    val deviceId : String
+)
