+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/model/response/LinkResponse.kt
@@ -1,0 +1,9 @@
+package com.linkedlist.linkllet.core.data.model.response
+
+import com.linkedlist.linkllet.core.data.model.Link
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LinkResponse(
+    val articleList : List<Link>
+)
