+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/remote/AuthRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.linkedlist.linkllet.core.data.source.remote
+
+import com.linkedlist.linkllet.core.data.model.Auth
+
+interface AuthRemoteDataSource {
+    suspend fun signUp() : Result<Auth>
+}
