+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/repository/AuthRepository.kt
@@ -1,0 +1,7 @@
+package com.linkedlist.linkllet.core.data.repository
+
+import kotlinx.coroutines.flow.Flow
+
+interface AuthRepository {
+    fun signUp() : Flow<Result<Boolean>>
+}
