+++ core/data/src/main/java/com/linkedlist/linkllet/core/data/repository/AuthRepositoryImpl.kt
@@ -1,0 +1,24 @@
+package com.linkedlist.linkllet.core.data.repository
+
+import com.linkedlist.linkllet.core.data.model.Auth.*
+import com.linkedlist.linkllet.core.data.source.remote.AuthRemoteDataSource
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import javax.inject.Inject
+
+class AuthRepositoryImpl @Inject constructor(
+    private val authRemoteDataSource: AuthRemoteDataSource
+) : AuthRepository {
+    override fun signUp(): Flow<Result<Boolean>> = flow {
+        authRemoteDataSource.signUp()
+            .onFailure {
+                emit(Result.failure(it))
+            }.onSuccess {
+                when (it) {
+                    SIGNED_UP -> emit(Result.success(true))
+                    ALREADY_SIGNED_UP -> emit(Result.success(true))
+                    FAILED -> emit(Result.success(false))
+                }
+            }
+    }
+}
