+++ feature/link/src/main/java/com/linkedlist/linkllet/feature/link/links/LinksViewModel.kt
@@ -5,7 +5,7 @@ import android.util.Log
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.linkedlist.linkellet.core.data.repository.LinkRepository
+import com.linkedlist.linkllet.core.data.repository.LinkRepository
 import com.linkedlist.linkllet.feature.link.navigation.FOLDER_ID
 import com.linkedlist.linkllet.feature.link.navigation.FOLDER_TITLE
 import dagger.hilt.android.lifecycle.HiltViewModel
